requirement.txt
streamlit
trulens-providers-cortex
PyPDF2
langchain
snowflake.core
snowflake-connector-python
bs4
pandas